--- a/src/elapi/api/__init__.py
+++ b/src/elapi/api/__init__.py
@@ -1,5 +1,6 @@
 # ruff: noqa: F401
 from .api import (
+    SessionDefaults,
     SimpleClient,
     GlobalSharedSession,
     APIRequest,