--- a/src/elapi/api/__init__.py
+++ b/src/elapi/api/__init__.py
@@ -1,6 +1,7 @@
 # ruff: noqa: F401
 from .api import (
-    CustomClient,
+    SimpleClient,
+    GlobalSharedSession,
     APIRequest,
     ElabFTWURLError,
     ElabFTWURL,