@mhxion 2024-09-04
# A few updates

Since the merge, a few commits related to this PR has been made to `dev` branch.

- elAPI will send a specific user-agent string `elAPI/<elAPI version> python-httpx/<HTTPX version>` (4bde6cd)
- Optional `uvloop` instead of `asyncio` for `bill-teams` plugin
- With an improved internal `asyncio` logic, `nested-asyncio` is no longer needed and has been removed